Sphere, MovingSphere to separate module
